GEOMESA-516 Factor out common code in IndexEntry and RasterIndexEntry
* Removed large amounts of duplicated code between IndexEntry and RasterIndexEntry.
* Moved most components into trait.
* Renamed RasterIndex to be RasterIndexEntry
