+++ geomesa-core/src/main/scala/org/locationtech/geomesa/core/index/IndexEntry.scala
@@ -16,66 +16,7 @@ import org.opengis.feature.simple.SimpleFeature
 
 import scala.collection.JavaConversions._
 
-object IndexEntry {
-
-  val timeZone = DateTimeZone.forID("UTC")
-
-  implicit class IndexEntrySFT(sf: SimpleFeature) {
-    lazy val userData = sf.getFeatureType.getUserData
-    lazy val dtgStartField = userData.getOrElse(SF_PROPERTY_START_TIME, DEFAULT_DTG_PROPERTY_NAME).asInstanceOf[String]
-    lazy val dtgEndField = userData.getOrElse(SF_PROPERTY_END_TIME, DEFAULT_DTG_END_PROPERTY_NAME).asInstanceOf[String]
-
-    lazy val sid = sf.getID
-    lazy val gh: GeoHash = GeohashUtils.reconstructGeohashFromGeometry(geometry)
-    def geometry = sf.getDefaultGeometry match {
-      case geo: Geometry => geo
-      case other =>
-        throw new Exception(s"Default geometry must be Geometry: '$other' of type '${Option(other).map(_.getClass).orNull}'")
-    }
-
-    private def getTime(attr: String) = sf.getAttribute(attr).asInstanceOf[java.util.Date]
-    def startTime = getTime(dtgStartField)
-    def endTime   = getTime(dtgEndField)
-    lazy val dt   = Option(startTime).map { d => new DateTime(d) }
-
-    private def setTime(attr: String, time: DateTime) =
-      sf.setAttribute(attr, Option(time).map(_.toDate).orNull)
-
-    def setStartTime(time: DateTime) = setTime(dtgStartField, time)
-    def setEndTime(time: DateTime)   = setTime(dtgEndField, time)
-  }
-
-  // the index value consists of the feature's:
-  // 1.  ID
-  // 2.  WKB-encoded geometry
-  // 3.  start-date/time
-  def encodeIndexValue(entry: SimpleFeature): Value = {
-    val encodedId = entry.sid.getBytes
-    val encodedGeom = WKBUtils.write(entry.geometry)
-    val encodedDtg = entry.dt.map(dtg => ByteBuffer.allocate(8).putLong(dtg.getMillis).array()).getOrElse(Array[Byte]())
-
-    new Value(
-               ByteBuffer.allocate(4).putInt(encodedId.length).array() ++ encodedId ++
-               ByteBuffer.allocate(4).putInt(encodedGeom.length).array() ++ encodedGeom ++
-               encodedDtg)
-  }
-
-  def decodeIndexValue(v: Value): DecodedIndexValue = {
-    val buf = v.get()
-    val idLength = ByteBuffer.wrap(buf, 0, 4).getInt
-    val (idPortion, geomDatePortion) = buf.drop(4).splitAt(idLength)
-    val id = new String(idPortion)
-    val geomLength = ByteBuffer.wrap(geomDatePortion, 0, 4).getInt
-    if(geomLength < (geomDatePortion.length - 4)) {
-      val (l,r) = geomDatePortion.drop(4).splitAt(geomLength)
-      DecodedIndexValue(id, WKBUtils.read(l), Some(ByteBuffer.wrap(r).getLong))
-    } else {
-      DecodedIndexValue(id, WKBUtils.read(geomDatePortion.drop(4)), None)
-    }
-  }
-
-  case class DecodedIndexValue(id: String, geom: Geometry, dtgMillis: Option[Long])
-}
+object IndexEntry extends IndexHelpers { }
 
 case class IndexEntryEncoder(rowf: TextFormatter,
                              cff: TextFormatter,
@@ -153,4 +94,69 @@ case class IndexEntryDecoder(ghDecoder: GeohashDecoder, dtDecoder: Option[DateDe
     builder.addAll(List(ghDecoder.decode(key).geom, dtDecoder.map(_.decode(key))))
     builder.buildFeature("")
   }
+}
+
+trait IndexHelpers {
+
+  val timeZone = DateTimeZone.forID("UTC")
+
+  implicit class IndexEntrySFT(sf: SimpleFeature) {
+    lazy val userData = sf.getFeatureType.getUserData
+    lazy val dtgStartField = userData.getOrElse(SF_PROPERTY_START_TIME, DEFAULT_DTG_PROPERTY_NAME).asInstanceOf[String]
+    lazy val dtgEndField = userData.getOrElse(SF_PROPERTY_END_TIME, DEFAULT_DTG_END_PROPERTY_NAME).asInstanceOf[String]
+
+    lazy val sid = sf.getID
+    lazy val gh: GeoHash = GeohashUtils.reconstructGeohashFromGeometry(geometry)
+    def geometry = sf.getDefaultGeometry match {
+      case geo: Geometry => geo
+      case other =>
+        throw new Exception(s"Default geometry must be Geometry: '$other' of type '${Option(other).map(_.getClass).orNull}'")
+    }
+
+    private def getTime(attr: String) = sf.getAttribute(attr).asInstanceOf[java.util.Date]
+    def startTime = getTime(dtgStartField)
+    def endTime   = getTime(dtgEndField)
+    lazy val dt   = Option(startTime).map { d => new DateTime(d) }
+
+    private def setTime(attr: String, time: DateTime) =
+      sf.setAttribute(attr, Option(time).map(_.toDate).orNull)
+
+    def setStartTime(time: DateTime) = setTime(dtgStartField, time)
+    def setEndTime(time: DateTime)   = setTime(dtgEndField, time)
+  }
+
+  def byteArrayToDecodedIndexValue(b: Array[Byte]): DecodedIndexValue = {
+    val idLength = ByteBuffer.wrap(b, 0, 4).getInt
+    val (idPortion, geomDatePortion) = b.drop(4).splitAt(idLength)
+    val id = new String(idPortion)
+    val geomLength = ByteBuffer.wrap(geomDatePortion, 0, 4).getInt
+    if(geomLength < (geomDatePortion.length - 4)) {
+      val (l,r) = geomDatePortion.drop(4).splitAt(geomLength)
+      DecodedIndexValue(id, WKBUtils.read(l), Some(ByteBuffer.wrap(r).getLong))
+    } else {
+      DecodedIndexValue(id, WKBUtils.read(geomDatePortion.drop(4)), None)
+    }
+  }
+
+  // the index value consists of the feature's:
+  // 1.  ID
+  // 2.  WKB-encoded geometry
+  // 3.  start-date/time
+  def encodeIndexValue(entry: SimpleFeature): Value = {
+    val encodedId = entry.sid.getBytes
+    val encodedGeom = WKBUtils.write(entry.geometry)
+    val encodedDtg = entry.dt.map(dtg => ByteBuffer.allocate(8).putLong(dtg.getMillis).array()).getOrElse(Array[Byte]())
+
+    new Value(
+      ByteBuffer.allocate(4).putInt(encodedId.length).array() ++ encodedId ++
+        ByteBuffer.allocate(4).putInt(encodedGeom.length).array() ++ encodedGeom ++
+        encodedDtg)
+  }
+
+  def decodeIndexValue(v: Value): DecodedIndexValue = {
+    val buf = v.get()
+    byteArrayToDecodedIndexValue(buf)
+  }
+
+  case class DecodedIndexValue(id: String, geom: Geometry, dtgMillis: Option[Long])
 }
